behavioral_dispositions/requirements.txt
google-genai
jinja2
pandas
tqdm
kagglehub
